refactor(backend): migrate all invalidate_release_snapshots call sites to stats_snapshot module
