+++ argus/backend/service/jira_service.py
@@ -12,7 +12,8 @@ from jira import JIRA
 
 from argus.backend.models.jira import JiraIssue
 from argus.backend.models.runtime_store import RuntimeStore
-from argus.backend.models.web import ArgusEventTypes, ArgusTest, ArgusUserView, invalidate_release_snapshots
+from argus.backend.models.web import ArgusEventTypes, ArgusTest, ArgusUserView
+from argus.backend.service.stats_snapshot import invalidate_release_snapshots
 from argus.backend.models.github_issue import IssueLink, IssueLabel
 from argus.backend.plugins.core import PluginInfoBase
 from argus.backend.plugins.loader import AVAILABLE_PLUGINS
